[BT-313]: Moved AppLogo in own subfolder and updated AppHeader example
