+++ src/framework/appHeader/_index.scss
@@ -1,7 +1,7 @@
 @import 'accountNav';
 @import 'accountPicture';
 @import 'appHeader';
-@import 'appLogo';
+@import 'logo/appLogo';
 @import 'appNav';
 @import 'appTitle';
 @import 'appTitleContainer';
