+++ src/framework/appHeader/logo/AppLogo.jsx
