+++ src/framework/appHeader/logo/AppLogo.spec.js
