+++ src/framework/appHeader/logo/_appLogo.scss
